Add notice everywhere PJs are displayed
